Fix version of `qlever` script and improve Docker image help message
As part of building the Docker image, the `qlever` command-line tool is
installed. So far, the version of this script was not fixed, which led
to random behavior due to caching issue. Now it is fixed to a version
that is compatible with the master.

On the side, improve the help message shown when running `docker run
adfreiburg/qlever` without arguments.