Handle blank nodes and local vocab entries correctly in the triple serializer (#2610)
So far, the serialization functions in `TripleSerializer.h` (which are used for persisting update triples in the `DeltaTriples` class) were not quite correct regarding blank nodes and local vocab entries. This is now improved in two ways:

1. The blank node blocks used by a `LocalVocab` are now correctly serialized and deserialized; see #2574
2. So far, only the `primaryWordSet` of a `LocalVocab` was serialized, now also the `otherWortSet`s are serialized

CAVEAT: This is much better than before but still not 100% correct in all cases. For example, consider the case where an update operation creates a new local vocab entry and an `Id` with a pointer to that local vocab entry becomes part of a query result. Such local vocab entries are not part of the local vocab of the query result. For example, when serializing this query result to disk, restarting the server, and reading the result back in again, the pointer encoded in that local vocab entry will be dangling.