Check whether to use internal permutation already in the query planner (#2613)
So far, the check whether the internal `PSO` or `POS` permutation should be used was made relatively late in the call stack, in the `Permutation` class. Now the decision is made in the query planner. There is a new free function `getPermutationForTriple`, which is defined and implemented in its own files `src/engine/PermutationSelector.{h,cpp}`

This change simplifies the code and also fixes the bug that currently only `Id`s holding a `VocabIndex` could be considered internal.

NOTE: This is good progress, but there is still work to do regarding the handling of internal permutations. In particular, the `Permutation` class still has a pointer `internalPermutation_` (which is a hack), and the decision which permutation to use should be made even earlier in the query planner. Eventually, the code should become oblivious to the nature of the permutation as early as possible